backend/requirements.txt
flask
flask-cors
openai-whisper
torch
deep-translator
gtts
